requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cleanup==5.1.0
django-filter==2.4.0
django-resized==0.3.11
djangorestframework==3.12.2
Pillow==8.0.1
pycryptodome==3.9.9
pytz==2020.4
sorl-thumbnail==12.7.0
sqlparse==0.4.1
